refactor: split roadmap banner preview data
